feat: Missing owner validation for deleting community

+++ src/controllers/handlers/delete-community-handler.ts
@@ -1,19 +1,35 @@
 import { HandlerContextWithPath, HTTPResponse } from '../../types'
 import { messageErrorOrUnknown } from '../../utils/errors'
 import { CommunityNotFoundError } from '../../adapters/errors'
+import { NotAuthorizedError } from '@dcl/platform-server-commons'
 
 export async function deleteCommunityHandler(
-  context: Pick<HandlerContextWithPath<'communitiesDb' | 'logs', '/communities/:id'>, 'components' | 'params'>
+  context: Pick<
+    HandlerContextWithPath<'communitiesDb' | 'logs', '/communities/:id'>,
+    'components' | 'params' | 'verification'
+  >
 ): Promise<HTTPResponse> {
   const {
     components: { communitiesDb, logs },
-    params: { id }
+    params: { id },
+    verification
   } = context
   const logger = logs.getLogger('privacy-handler')
-
   logger.info(`Deleting community: ${id}`)
 
+  const userAddress = verification?.auth.toLowerCase()
+
+  if (!userAddress) {
+    throw new NotAuthorizedError('Unauthorized')
+  }
+
   try {
+    const community = await communitiesDb.getCommunity(id, userAddress)
+
+    if (community.ownerAddress.toLowerCase() !== userAddress) {
+      throw new NotAuthorizedError("The user doesn't have permission to delete this community")
+    }
+
     await communitiesDb.deleteCommunity(id)
 
     return {
@@ -23,6 +39,10 @@ export async function deleteCommunityHandler(
     logger.error(`Error getting community: ${id}, error: ${messageErrorOrUnknown(error)}`)
 
     if (error instanceof CommunityNotFoundError) {
+      throw error
+    }
+
+    if (error instanceof NotAuthorizedError) {
       throw error
     }
